remove SysFileService prototype
